+++ android/src/main/java/com/bitmovin/player/reactnative/RNPlayerViewManager.kt
@@ -115,6 +115,7 @@ class RNPlayerViewManager : Module() {
                 "onBmpSeeked",
                 "onBmpTimeShift",
                 "onBmpTimeShifted",
+                "onBmpDvrWindowExceeded",
                 "onBmpStallStarted",
                 "onBmpStallEnded",
                 "onBmpTimeChanged",