+++ ios/RNPlayerViewManager.swift
@@ -52,6 +52,7 @@ public class RNPlayerViewManager: Module {
                 "onBmpSeeked",
                 "onBmpTimeShift",
                 "onBmpTimeShifted",
+                "onBmpDvrWindowExceeded",
                 "onBmpStallStarted",
                 "onBmpStallEnded",
                 "onBmpTimeChanged",